requirements.txt
click>=8.1.0
requests>=2.31.0
tabulate>=0.9.0
python-dotenv>=1.0.0
